made an almost done ui
players can select and press button but the buttons doesnt do anything and when a button is selected, it just changes the text.
But fundamentals almost done.
