--- Project Bergamot/Assets/Controllers/Input/InputSystem.inputactions.meta
@@ -1,14 +1,0 @@
-fileFormatVersion: 2
-guid: b8dd28fe7a1e1404e9cfd301bc2d6123
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
-  generateWrapperCode: 0
-  wrapperCodePath: 
-  wrapperClassName: 
-  wrapperCodeNamespace:
