--- Project Bergamot/Assets/Marked.cs
@@ -1,23 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Marked : MonoBehaviour
-{
-    public static List<GameObject> playerObject = new List<GameObject>();
-
-    public List<string> playerName2 = new List<string>();
-
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-    }
-}
