--- Project Bergamot/Assets/Scripts/InputSystem.inputsettings.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 60594440bd9317644a53f2b49b9398a2
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
