+++ Project Bergamot/Assets/Scripts/UI/PlayerScript.cs.meta
